refactor: 패키지명 변경 (service -> application)
