+++ src/main/java/feedzupzup/feedzupzupmanager/global/config/AiToolConfig.java
@@ -3,8 +3,8 @@ package feedzupzup.feedzupzupmanager.global.config;
 import static feedzupzup.feedzupzupmanager.ai.constant.AiPrompts.DBA_SYSTEM_PROMPT;
 
 import feedzupzup.feedzupzupmanager.ai.util.AiTaskWrapper;
-import feedzupzup.feedzupzupmanager.ingest.service.VectorStoreAdapter;
-import feedzupzup.feedzupzupmanager.query.service.QueryService;
+import feedzupzup.feedzupzupmanager.ingest.application.VectorStoreAdapter;
+import feedzupzup.feedzupzupmanager.query.application.QueryService;
 import java.util.function.Function;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.ai.chat.client.ChatClient;
