+++ src/main/java/feedzupzup/feedzupzupmanager/ingest/application/SwaggerGateway.java
@@ -1,4 +1,4 @@
-package feedzupzup.feedzupzupmanager.ingest.service;
+package feedzupzup.feedzupzupmanager.ingest.application;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
