+++ src/main/java/feedzupzup/feedzupzupmanager/query/application/QueryService.java
@@ -1,4 +1,4 @@
-package feedzupzup.feedzupzupmanager.query.service;
+package feedzupzup.feedzupzupmanager.query.application;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
